// packages/twenty-server/src/engine/metadata-modules/command-menu-item/enums/engine-component-key.enum.ts
export enum EngineComponentKey {
  NAVIGATE_TO_NEXT_RECORD = 'NAVIGATE_TO_NEXT_RECORD',
  NAVIGATE_TO_PREVIOUS_RECORD = 'NAVIGATE_TO_PREVIOUS_RECORD',
  CREATE_NEW_RECORD = 'CREATE_NEW_RECORD',
  DELETE_RECORDS = 'DELETE_RECORDS',
  RESTORE_RECORDS = 'RESTORE_RECORDS',
  DESTROY_RECORDS = 'DESTROY_RECORDS',
  ADD_TO_FAVORITES = 'ADD_TO_FAVORITES',
  REMOVE_FROM_FAVORITES = 'REMOVE_FROM_FAVORITES',
  EXPORT_NOTE_TO_PDF = 'EXPORT_NOTE_TO_PDF',
  EXPORT_RECORDS = 'EXPORT_RECORDS',
  UPDATE_MULTIPLE_RECORDS = 'UPDATE_MULTIPLE_RECORDS',
  MERGE_MULTIPLE_RECORDS = 'MERGE_MULTIPLE_RECORDS',
  IMPORT_RECORDS = 'IMPORT_RECORDS',
  EXPORT_VIEW = 'EXPORT_VIEW',
  SEE_DELETED_RECORDS = 'SEE_DELETED_RECORDS',
  CREATE_NEW_VIEW = 'CREATE_NEW_VIEW',
  HIDE_DELETED_RECORDS = 'HIDE_DELETED_RECORDS',
  EDIT_RECORD_PAGE_LAYOUT = 'EDIT_RECORD_PAGE_LAYOUT',
  EDIT_DASHBOARD_LAYOUT = 'EDIT_DASHBOARD_LAYOUT',
  SAVE_DASHBOARD_LAYOUT = 'SAVE_DASHBOARD_LAYOUT',
  CANCEL_DASHBOARD_LAYOUT = 'CANCEL_DASHBOARD_LAYOUT',
  DUPLICATE_DASHBOARD = 'DUPLICATE_DASHBOARD',
  DUPLICATE_MESSAGE_LIST = 'DUPLICATE_MESSAGE_LIST',
  ACTIVATE_WORKFLOW = 'ACTIVATE_WORKFLOW',
  DEACTIVATE_WORKFLOW = 'DEACTIVATE_WORKFLOW',
  DISCARD_DRAFT_WORKFLOW = 'DISCARD_DRAFT_WORKFLOW',
  TEST_WORKFLOW = 'TEST_WORKFLOW',
  SEE_ACTIVE_VERSION_WORKFLOW = 'SEE_ACTIVE_VERSION_WORKFLOW',
  SEE_RUNS_WORKFLOW = 'SEE_RUNS_WORKFLOW',
  SEE_VERSIONS_WORKFLOW = 'SEE_VERSIONS_WORKFLOW',
  ADD_NODE_WORKFLOW = 'ADD_NODE_WORKFLOW',
  TIDY_UP_WORKFLOW = 'TIDY_UP_WORKFLOW',
  DUPLICATE_WORKFLOW = 'DUPLICATE_WORKFLOW',
  TOGGLE_WORKFLOW_VISIBILITY = 'TOGGLE_WORKFLOW_VISIBILITY',
  SEE_VERSION_WORKFLOW_RUN = 'SEE_VERSION_WORKFLOW_RUN',
  SEE_WORKFLOW_WORKFLOW_RUN = 'SEE_WORKFLOW_WORKFLOW_RUN',
  STOP_WORKFLOW_RUN = 'STOP_WORKFLOW_RUN',
  RETRY_WORKFLOW_RUN = 'RETRY_WORKFLOW_RUN',
  SEE_RUNS_WORKFLOW_VERSION = 'SEE_RUNS_WORKFLOW_VERSION',
  SEE_WORKFLOW_WORKFLOW_VERSION = 'SEE_WORKFLOW_WORKFLOW_VERSION',
  USE_AS_DRAFT_WORKFLOW_VERSION = 'USE_AS_DRAFT_WORKFLOW_VERSION',
  SEE_VERSIONS_WORKFLOW_VERSION = 'SEE_VERSIONS_WORKFLOW_VERSION',
  SEARCH_RECORDS = 'SEARCH_RECORDS',
  SEARCH_RECORDS_FALLBACK = 'SEARCH_RECORDS_FALLBACK',
  ASK_AI = 'ASK_AI',
  VIEW_PREVIOUS_AI_CHATS = 'VIEW_PREVIOUS_AI_CHATS',
  NEW_AI_CHAT = 'NEW_AI_CHAT',
  SHARE_RECORD = 'SHARE_RECORD',
  NAVIGATION = 'NAVIGATION',
  TRIGGER_WORKFLOW_VERSION = 'TRIGGER_WORKFLOW_VERSION',
  FRONT_COMPONENT_RENDERER = 'FRONT_COMPONENT_RENDERER',
  REPLY_TO_EMAIL_THREAD = 'REPLY_TO_EMAIL_THREAD',
  COMPOSE_EMAIL = 'COMPOSE_EMAIL',
  COMPOSE_EMAIL_TO_RELATED_PEOPLE = 'COMPOSE_EMAIL_TO_RELATED_PEOPLE',
  COMPOSE_CAMPAIGN = 'COMPOSE_CAMPAIGN',
  SEND_MESSAGE_CAMPAIGN = 'SEND_MESSAGE_CAMPAIGN',
  SEND_MESSAGE_CAMPAIGN_TEST = 'SEND_MESSAGE_CAMPAIGN_TEST',
  CANCEL_MESSAGE_CAMPAIGN = 'CANCEL_MESSAGE_CAMPAIGN',
  DUPLICATE_MESSAGE_CAMPAIGN = 'DUPLICATE_MESSAGE_CAMPAIGN',
  EMAIL_BLOCK_SETTINGS = 'EMAIL_BLOCK_SETTINGS',

  // TODO: Remove deprecated keys once upgrade:1-21:refactor-navigation-commands has run on all workspaces
  // Deprecated: replaced by NAVIGATION engine key with payload
  GO_TO_PEOPLE = 'GO_TO_PEOPLE',
  GO_TO_COMPANIES = 'GO_TO_COMPANIES',
  GO_TO_DASHBOARDS = 'GO_TO_DASHBOARDS',
  GO_TO_OPPORTUNITIES = 'GO_TO_OPPORTUNITIES',
  GO_TO_SETTINGS = 'GO_TO_SETTINGS',
  GO_TO_TASKS = 'GO_TO_TASKS',
  GO_TO_NOTES = 'GO_TO_NOTES',
  GO_TO_WORKFLOWS = 'GO_TO_WORKFLOWS',
  GO_TO_RUNS = 'GO_TO_RUNS',

  // Deprecated keys kept for backward compatibility until migration runs
  DELETE_SINGLE_RECORD = 'DELETE_SINGLE_RECORD',
  DELETE_MULTIPLE_RECORDS = 'DELETE_MULTIPLE_RECORDS',
  RESTORE_SINGLE_RECORD = 'RESTORE_SINGLE_RECORD',
  RESTORE_MULTIPLE_RECORDS = 'RESTORE_MULTIPLE_RECORDS',
  DESTROY_SINGLE_RECORD = 'DESTROY_SINGLE_RECORD',
  DESTROY_MULTIPLE_RECORDS = 'DESTROY_MULTIPLE_RECORDS',
  EXPORT_FROM_RECORD_INDEX = 'EXPORT_FROM_RECORD_INDEX',
  EXPORT_FROM_RECORD_SHOW = 'EXPORT_FROM_RECORD_SHOW',
  EXPORT_MULTIPLE_RECORDS = 'EXPORT_MULTIPLE_RECORDS',
}